+++ UI_ELEMENTS/WINDOWS/TIMER_WINDOW/Element.lua
@@ -522,14 +522,14 @@ function TimerWindowElement:SortChildren()
 
                 -- both permanent > sort by index > child2 first
                 if child1.data.permanent == true and
-                child1.data.permanent == true and
+                child2.data.permanent == true and
                 child1.data.sortIndex > child2.data.sortIndex then
 
                     return true
 
                 -- both permanent > sort by index > child1 first
                 elseif child1.data.permanent == true and
-                child1.data.permanent == true and
+                child2.data.permanent == true and
                 child1.data.sortIndex < child2.data.sortIndex then
 
                     return false
@@ -566,14 +566,14 @@ function TimerWindowElement:SortChildren()
 
                 -- both permanent > sort by index > child2 first
                 if child1.data.permanent == true and
-                child1.data.permanent == true and
+                child2.data.permanent == true and
                 child1.data.sortIndex > child2.data.sortIndex then
 
                     return false
 
                 -- both permanent > sort by index > child1 first
                 elseif child1.data.permanent == true and
-                child1.data.permanent == true and
+                child2.data.permanent == true and
                 child1.data.sortIndex < child2.data.sortIndex then
 
                     return true